requirements.txt
ccxt
aiohttp
fastapi
uvicorn
pydantic
python-dotenv
PyYAML
numpy
pandas
scipy
prometheus-client
jinja2
pytest
ruff
flake8
xgboost
scikit-learn
joblib
openai

# Fine-tuned LLM dependencies
transformers>=4.35.0
torch>=2.0.0
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
sentencepiece
protobuf

# RAG dependencies
langchain>=0.1.0
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA
chromadb>=0.4.0
pypdf>=3.0.0
tiktoken

# Embeddings
sentence-transformers>=2.2.0

# Security
guardrails-ai>=0.4.0

# Monitoring
wandb>=0.16.0
mlflow>=2.8.0

# UI
gradio>=4.0.0
streamlit>=1.28.0

# Voice/Transcription
openai-whisper>=20231117

# Caching
redis>=5.0.0

# Rate limiting
slowapi>=0.1.9

# Additional utilities
requests>=2.31.0

# Phase 1: Multi-Exchange & WebSocket Support
websockets>=12.0  # WebSocket support for real-time data
cryptography>=41.0.0  # API key encryption

# Phase 2+: Complete Overhaul Dependencies
# Machine Learning
tensorflow>=2.12.0  # LSTM price prediction
keras>=2.12.0  # Deep learning framework

# Backtesting
backtrader>=1.9.76.123  # Backtesting engine

# Telegram Alerts
python-telegram-bot>=20.3  # Telegram notifications

# Tax Reporting (optional)
# koinly>=0.1.5  # Uncomment if using Koinly API

# Genetic Algorithm Optimization (optional)
deap>=1.3.3  # Genetic algorithms for strategy optimization

# Advanced HFT Features
cvxpy>=1.3.0  # Portfolio optimization
feedparser>=6.0.0  # RSS feed parsing
aiohttp>=3.9.0  # Async HTTP client
tweepy>=4.14.0  # Twitter API (optional)
praw>=7.7.0  # Reddit API (optional)
transformers>=4.35.0  # Sentiment analysis (already included but ensuring version)
streamlit>=1.28.0  # Advanced dashboard (already included)
plotly>=5.17.0  # Interactive charts